Implemented authorization to the update and delete task fucntion

+++ src/tasks/controller.py
@@ -19,8 +19,8 @@ def create_task(body: TaskSchema, db: Session,user:UserModel):
     return new_task
 
 
-def get_tasks(db: Session):
-    tasks = db.query(TaskModel).all()
+def get_tasks(db: Session,user:UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id==user.id).all()
     return tasks
 
 
@@ -32,12 +32,14 @@ def get_one_task(task_id: int, db: Session):
     return one_task
 
 
-def update_task(body: TaskSchema, task_id: int, db: Session):
+def update_task(body: TaskSchema, task_id: int, db: Session,user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
-
     if not one_task:
         raise HTTPException(404, {"details": "Task not found"})
 
+    if one_task.user_id !=user.id:
+        raise HTTPException(404, detail="You are not authorized to edit this task")
+    
     body = body.model_dump()
     for field, value in body.items():
         setattr(one_task, field, value)
@@ -49,11 +51,14 @@ def update_task(body: TaskSchema, task_id: int, db: Session):
     return one_task
 
 
-def delete_task(task_id: int, db: Session):
+def delete_task(task_id: int, db: Session,user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, {"details": "Task not found"})
 
+    if one_task.user_id !=user.id:
+            raise HTTPException(404, detail="You are not authorized to delete this task")
+        
     db.delete(one_task)
     db.commit()
